moving XR to ./src/
