+++ src/Cameras/VR/vrExperienceHelper.ts
@@ -36,9 +36,9 @@ import "../../Meshes/Builders/cylinderBuilder";
 import "../../Gamepads/gamepadSceneComponent";
 import "../../Animations/animatable";
 import { Axis } from '../../Maths/math.axis';
-import { WebXRSessionManager } from '../XR/webXRSessionManager';
-import { WebXRDefaultExperience } from '../XR/webXRDefaultExperience';
-import { WebXRState } from '../XR/webXRTypes';
+import { WebXRSessionManager } from '../../XR/webXRSessionManager';
+import { WebXRDefaultExperience } from '../../XR/webXRDefaultExperience';
+import { WebXRState } from '../../XR/webXRTypes';
 
 /**
  * Options to modify the vr teleportation behavior.
