+++ src/Cameras/index.ts
@@ -17,5 +17,4 @@ export * from "./Stereoscopic/index";
 export * from "./universalCamera";
 export * from "./virtualJoysticksCamera";
 export * from "./VR/index";
-export * from "./XR/index";
 export * from "./RigModes/index";
