+++ src/XR/features/WebXRPlaneDetector.ts
@@ -1,8 +1,8 @@
 import { WebXRFeaturesManager, WebXRFeatureName } from '../webXRFeaturesManager';
-import { TransformNode } from '../../../Meshes/transformNode';
+import { TransformNode } from '../../Meshes/transformNode';
 import { WebXRSessionManager } from '../webXRSessionManager';
-import { Observable } from '../../../Misc/observable';
-import { Vector3, Matrix } from '../../../Maths/math.vector';
+import { Observable } from '../../Misc/observable';
+import { Vector3, Matrix } from '../../Maths/math.vector';
 import { WebXRAbstractFeature } from './WebXRAbstractFeature';
 
 /**
