+++ src/index.ts
@@ -36,4 +36,5 @@ export * from "./sceneComponent";
 export * from "./Sprites/index";
 export * from "./States/index";
 export * from "./Misc/index";
+export * from "./XR/index";
 export * from "./types";
